requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.8.2
blinker==1.6.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
cytoolz==0.12.2
dbus-python==1.2.18
discord.py==2.3.2
distro==1.7.0
distro-info===1.1build1
et-xmlfile==1.1.0
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
Flask==3.0.0
frozenlist==1.4.0
hexbytes==0.3.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lru-dict==1.2.0
MarkupSafe==2.1.3
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
openpyxl==3.1.2
pandas==2.0.2
parsimonious==0.9.0
protobuf==4.24.4
pycryptodome==3.19.0
PyGObject==3.42.1
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
pytz==2023.3
pyunormalize==15.0.0
PyYAML==5.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.6
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
toolz==0.12.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
wadllib==1.3.6
web3==6.11.1
websockets==11.0.3
Werkzeug==3.0.0
yarl==1.9.2
zipp==1.0.0
